library_system/gateway/requirements.txt
Django==4.1.1
djangorestframework==3.14.0
requests==2.25.1
django-cors-headers
pyjwt
cryptography